+++ idea/testData/refactoring/introduceVariable/skipPackageReference.kt
@@ -1,0 +1,6 @@
+// WITH_RUNTIME
+import java.util.*
+
+fun foo() {
+    <caret>Arrays.asList(1, 2, 3)
+}