+++ idea/testData/smartSelection/simple.kt
@@ -5,5 +5,4 @@ fun main(args: Array<String>) {
 1.0
 1.0 + 1
 Math.pow(...)
-print(...)
 */